desafio2/requirements.txt
flask
pymysql
pymongo
python-dotenv